:recycle: refactor(README): Update README and add requirements
Updated README.md and new packages added in requirements.txt. The change in README.md is a minor stylistic update for easier readability. The requirements.txt file has been updated to include Django, python-dateutil, Faker, and setuptools. This is to ensure the smooth running of the app with the current versions of dependencies being used.